getPartyNamesForUser() function fixed.
Need to be double checked if it actually adds the info to the Users object being called on, however it does get the data and attempt to append as far as I've seen.
loadReservations() will be completed next, considering this function needed the list of parties for the user to operate.